Make the code more OOP - Add UI class to deal with interactions with the user - Add Storage class to deal with loading/saving tasks in file - Add Parser class to deal with user commands - Add TaskList class to handle operations relating to tasks
