+++ src/main/java/Deadline.java
@@ -1,7 +1,7 @@
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
 
     protected LocalDateTime by;
